Extract logic for aggregating body analysis results and generating endpoint bodies to a projection
