+++ workspaces/diff-engine/cli/src/learn.rs
@@ -11,10 +11,12 @@ use tokio::io::{stdin, stdout, AsyncWrite};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::ReceiverStream;
 
+use optic_diff_engine::streams;
 use optic_diff_engine::streams::{TaggedInput, Tags};
+use optic_diff_engine::Aggregate;
 use optic_diff_engine::{
-  analyze_documented_bodies, analyze_undocumented_bodies, streams, EndpointCommand,
-  InteractionDiffResult, SpecCommand,
+  analyze_documented_bodies, analyze_undocumented_bodies, EndpointCommand, InteractionDiffResult,
+  LearnedUndocumentedBodiesProjection, SpecCommand,
 };
 use optic_diff_engine::{
   BodyAnalysisLocation, HttpInteraction, JsonTrail, SpecChunkEvent, SpecEvent, SpecIdGenerator,
@@ -146,40 +148,17 @@ async fn learn_undocumented_bodies<S: 'static + AsyncWrite + Unpin + Send>(
     let mut analysiss = ReceiverStream::new(analysis_receiver);
     let mut id_generator = IdGenerator::default();
 
-    let mut observations_by_body_location = HashMap::new();
+    let mut learned_undocumented_bodies = LearnedUndocumentedBodiesProjection::default();
+
     while let Some(analysis) = analysiss.next().await {
-      let existing_observations = observations_by_body_location
-        .entry(analysis.body_location)
-        .or_insert_with(|| TrailObservationsResult::default());
-
-      existing_observations.union(analysis.trail_observations);
+      learned_undocumented_bodies.apply(analysis);
     }
 
-    let mut endpoints_by_endpoint = HashMap::new();
-    for (body_location, observations) in observations_by_body_location {
-      let (root_shape_id, body_commands) = observations.into_commands(&mut id_generator);
-      let endpoint_body = EndpointBody::new(&body_location, root_shape_id, body_commands);
+    let endpoint_bodies = learned_undocumented_bodies
+      .into_endpoint_bodies(&mut id_generator)
+      .collect::<Vec<_>>();
 
-      let (path_id, method) = match body_location {
-        BodyAnalysisLocation::UnmatchedRequest {
-          path_id, method, ..
-        } => (path_id, method),
-        BodyAnalysisLocation::UnmatchedResponse {
-          path_id, method, ..
-        } => (path_id, method),
-        _ => unreachable!("analyzing undocumented bodies should only yield unmatched results"),
-      };
-
-      let endpoint_bodies = endpoints_by_endpoint
-        .entry((path_id, method))
-        .or_insert_with_key(|(path_id, method)| {
-          EndpointBodies::new(path_id.clone(), method.clone())
-        });
-
-      endpoint_bodies.push(endpoint_body);
-    }
-
-    streams::write_to_json_lines(sink, endpoints_by_endpoint.values())
+    streams::write_to_json_lines(sink, endpoint_bodies.iter())
       .await
       .expect("could not write endpoint bodies to stdout");
   });
@@ -341,179 +320,6 @@ impl SpecIdGenerator for IdGenerator {
     // - 17 years for a 1% chance of at least one global collision assuming
     //   writing 1000 ids per hour (https://zelark.github.io/nano-id-cc/)
     format!("{}{}", prefix, nanoid!(10))
-  }
-}
-
-// Output types: undocumented bodies
-// ---------------------------------
-
-#[derive(Default, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct EndpointBodies {
-  path_id: String,
-  method: String,
-  requests: Vec<EndpointRequestBody>,
-  responses: Vec<EndpointResponseBody>,
-}
-
-impl EndpointBodies {
-  pub fn new(path_id: String, method: String) -> Self {
-    Self {
-      path_id,
-      method,
-      requests: vec![],
-      responses: vec![],
-    }
-  }
-
-  pub fn push(&mut self, endpoint: EndpointBody) {
-    match endpoint {
-      EndpointBody::Request(endpoint_request) => {
-        self.requests.push(endpoint_request);
-      }
-      EndpointBody::Response(endpoint_response) => {
-        self.responses.push(endpoint_response);
-      }
-    }
-  }
-}
-
-#[derive(Debug)]
-enum EndpointBody {
-  Request(EndpointRequestBody),
-  Response(EndpointResponseBody),
-}
-
-#[derive(Default, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct EndpointRequestBody {
-  commands: Vec<SpecCommand>,
-
-  #[serde(skip)]
-  path_id: String,
-  #[serde(skip)]
-  method: String,
-
-  #[serde(flatten)]
-  body_descriptor: Option<EndpointBodyDescriptor>,
-}
-
-#[derive(Default, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct EndpointResponseBody {
-  commands: Vec<SpecCommand>,
-  status_code: u16,
-
-  #[serde(skip)]
-  path_id: String,
-  #[serde(skip)]
-  method: String,
-
-  #[serde(flatten)]
-  body_descriptor: Option<EndpointBodyDescriptor>,
-}
-
-#[derive(Default, Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct EndpointBodyDescriptor {
-  content_type: String,
-  root_shape_id: String,
-}
-
-impl EndpointBody {
-  fn new(
-    body_location: &BodyAnalysisLocation,
-    root_shape_id: Option<String>,
-    body_commands: impl IntoIterator<Item = SpecCommand>,
-  ) -> Self {
-    let body_descriptor = match root_shape_id {
-      Some(root_shape_id) => Some(EndpointBodyDescriptor {
-        content_type: body_location
-          .content_type()
-          .expect("root shape id implies a content type to be present")
-          .clone(),
-        root_shape_id,
-      }),
-      None => None,
-    };
-
-    let mut body = match body_location {
-      BodyAnalysisLocation::UnmatchedRequest {
-        path_id, method, ..
-      } => EndpointBody::Request(EndpointRequestBody {
-        body_descriptor,
-        path_id: path_id.clone(),
-        method: method.clone(),
-        commands: body_commands.into_iter().collect(),
-      }),
-      BodyAnalysisLocation::UnmatchedResponse {
-        status_code,
-        path_id,
-        method,
-        ..
-      } => EndpointBody::Response(EndpointResponseBody {
-        body_descriptor,
-        path_id: path_id.clone(),
-        method: method.clone(),
-        commands: body_commands.into_iter().collect(),
-        status_code: *status_code,
-      }),
-      _ => panic!("EndpointBody should only be created for unmatched responses and requests"),
-    };
-
-    body.append_endpoint_commands();
-
-    body
-  }
-
-  fn append_endpoint_commands(&mut self) {
-    let mut ids = IdGenerator::default();
-
-    match self {
-      EndpointBody::Request(request_body) => {
-        let request_id = ids.request();
-        request_body
-          .commands
-          .push(SpecCommand::from(EndpointCommand::add_request(
-            request_id.clone(),
-            request_body.path_id.clone(),
-            request_body.method.clone(),
-          )));
-
-        if let Some(body_descriptor) = &request_body.body_descriptor {
-          request_body
-            .commands
-            .push(SpecCommand::from(EndpointCommand::set_request_body_shape(
-              request_id,
-              body_descriptor.root_shape_id.clone(),
-              body_descriptor.content_type.clone(),
-              false,
-            )));
-        }
-      }
-      EndpointBody::Response(response_body) => {
-        let response_id = ids.response();
-        response_body.commands.push(SpecCommand::from(
-          EndpointCommand::add_response_by_path_and_method(
-            response_id.clone(),
-            response_body.path_id.clone(),
-            response_body.method.clone(),
-            response_body.status_code.clone(),
-          ),
-        ));
-
-        if let Some(body_descriptor) = &response_body.body_descriptor {
-          response_body
-            .commands
-            .push(SpecCommand::from(EndpointCommand::set_response_body_shape(
-              response_id,
-              body_descriptor.root_shape_id.clone(),
-              body_descriptor.content_type.clone(),
-              false,
-            )));
-        }
-      }
-    };
   }
 }
