+++ workspaces/diff-engine/src/projections/mod.rs
@@ -1,6 +1,7 @@
 pub mod conflicts;
 pub mod endpoint;
 pub mod history;
+pub mod learners;
 pub mod shape;
 pub mod spec_events;
 pub mod spectacle;
@@ -8,6 +9,7 @@ pub mod spectacle;
 pub use conflicts::ConflictsProjection;
 pub use endpoint::EndpointProjection;
 pub use history::{CommitId, HistoryProjection};
+pub use learners::undocumented_bodies::LearnedUndocumentedBodiesProjection;
 pub use shape::ShapeProjection;
 pub use spec_events::{SpecAssemblerError, SpecAssemblerProjection};
 pub use spectacle::endpoints::EndpointsProjection;
